fix: next-auth option 분리 및 타입 수정
